# public/openapi.yaml
openapi: 3.1.0
info:
  title: BasaltSurge API
  version: 1.0.0
  contact:
    name: BasaltSurge Support
    url: https://surge.basalthq.com
    email: support@basalthq.com
  description: >
    Public API surface for BasaltSurge under a hardened posture: Azure Front
    Door Premium (AFD) gateway → Azure API Management (APIM).

    All developer-facing endpoints require an APIM subscription key header `Ocp-Apim-Subscription-Key` and use the APIM custom domain. AFD is optional as a fallback.

    APIM policy permits requests with a valid subscription key and may also accept AFD-originated requests carrying `x-edge-secret` if AFD is configured. Direct backend origins are not supported.

    Wallet identity is resolved at the gateway based on subscription; client-supplied `x-wallet` header and `wallet` parameters are ignored.

    Health endpoint: `GET /healthz` returns 200 JSON for probes; does not require a subscription.

    Note: Some endpoints are read-only or environment-gated in production. Admin/sensitive routes (JWT-only) are not part of this spec.
  x-guidance: >
    Welcome to BasaltSurge's Agentic Commerce API.

    To use this API:

    1. Non-paid resources are fully accessible with standard API keys or without auth depending on the path.

    2. Payment-gated resources under /api/x402/ must be paid for. For example, POST /api/x402/orders to discover the pricing required. The server will reply with HTTP 402 and an x402 payment challenge.

...

...
  x-internal-headers:
...
servers:
...
security:
  - apimKey: []
...
paths:
...
  /api/receipts:
    get:
      summary: List receipts
      operationId: listReceipts
      tags:
        - Receipts
      description: >-
        List recent receipts for the merchant associated with the resolved
        wallet. Requires APIM subscription and `receipts:read` scope. Gracefully
        degrades to in-memory when Cosmos is unavailable.



        Requires API key. Agents must use /api/x402/subscribe to pay and provision a key.
      x-required-scopes:
        - receipts:read
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            format: int32
            minimum: 1
            maximum: 200
            default: 100
          required: false
          description: Number of receipts to return (default 100, max 200)
      responses:
        "200":
          description: Recent receipts
          headers:
            X-RateLimit-Limit:
              $ref: "#/components/headers/X-RateLimit-Limit"
            X-RateLimit-Remaining:
              $ref: "#/components/headers/X-RateLimit-Remaining"
            X-RateLimit-Reset:
              $ref: "#/components/headers/X-RateLimit-Reset"
          content:
            application/json:
              schema:
                type: object
                properties:
                  receipts:
                    type: array
                    items:
                      $ref: "#/components/schemas/Receipt"
              examples:
                default:
                  value:
                    receipts:
                      - receiptId: R-123456
                        totalUsd: 13.09
                        currency: USD
                        lineItems:
                          - label: Espresso
                            priceUsd: 7
                            qty: 2
                          - label: Tax
                            priceUsd: 1.09
                          - label: Processing Fee
                            priceUsd: 0.5
                        createdAt: 1698765432000
                        brandName: BasaltSurge
                        status: paid
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "429":
          $ref: "#/components/responses/TooManyRequests"
      security:
        - apimKey: []
    post:
      summary: Create a receipt payload for a QR-code payment portal
      operationId: createReceipt
      tags:
        - Receipts
      description: >-
        Create an itemized receipt used by the payment portal. The portal URL
        can be QR-encoded and printed on POS receipts. Requires APIM
        subscription and `receipts:write` scope. Requests must traverse AFD;
        direct APIM origin may be rejected with 403 (origin enforcement).



        Requires API key. Agents must use /api/x402/subscribe to pay and provision a key.
      x-required-scopes:
        - receipts:write
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReceiptRequest"
            examples:
              default:
                value:
                  id: rcpt_12345
                  lineItems:
                    - label: Sample Item
                      priceUsd: 25
                    - label: Tax
                      priceUsd: 2
                  totalUsd: 27
      responses:
        "201":
          description: Receipt created
          headers:
            X-RateLimit-Limit:
              $ref: "#/components/headers/X-RateLimit-Limit"
            X-RateLimit-Remaining:
              $ref: "#/components/headers/X-RateLimit-Remaining"
            X-RateLimit-Reset:
              $ref: "#/components/headers/X-RateLimit-Reset"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateReceiptResponse"
              examples:
                default:
                  value:
                    id: rcpt_12345
                    paymentUrl: https://pay.ledger1.ai/pay/rcpt_12345
                    status: pending
        "400":
          description: Invalid request
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          description: Server error
      security:
        - apimKey: []
...
  /api/receipts/status:
    get:
      summary: Get receipt/payment status
      operationId: getReceiptStatus
      tags:
        - Receipts
      description: >-
        Requires APIM subscription and `receipts:read` scope. Requests must
        traverse AFD; direct APIM origin may be rejected with 403 (origin
        enforcement).



        Requires API key. Agents must use /api/x402/subscribe to pay and provision a key.
      x-required-scopes:
        - receipts:read
      parameters:
        - in: query
          name: receiptId
          schema:
            type: string
          required: true
          description: Receipt ID to check
      responses:
        "200":
          description: Status payload
          headers:
            X-RateLimit-Limit:
              $ref: "#/components/headers/X-RateLimit-Limit"
            X-RateLimit-Remaining:
              $ref: "#/components/headers/X-RateLimit-Remaining"
            X-RateLimit-Reset:
              $ref: "#/components/headers/X-RateLimit-Reset"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReceiptStatus"
              examples:
                default:
                  value:
                    id: rcpt_12345
                    status: completed
                    transactionHash: 0xabc123...
                    currency: USDC
                    amount: 27
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "404":
          description: Not found
        "429":
          $ref: "#/components/responses/TooManyRequests"
      security:
        - apimKey: []
    post:
      summary: Update receipt status (tracking and sensitive events)
      operationId: postReceiptStatus
      tags:
        - Receipts
      description: >-
        Updates receipt status timeline. Certain tracking statuses may be
        allowed without JWT (e.g. link_opened, buyer_logged_in,
        checkout_initialized). Sensitive updates (e.g. checkout_success,
        refund_*) require JWT and CSRF via the portal UI; APIM developer
        subscriptions typically use tracking statuses.



        Requires API key. Agents must use /api/x402/subscribe to pay and provision a key.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReceiptStatusUpdateRequest"
      responses:
        "200":
          description: Status update result
          headers:
            x-correlation-id:
              description: Correlation ID for tracing
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
        "400":
          description: missing_receipt_id or invalid_wallet or missing_status
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          description: failed
      security:
        - apimKey: []
  /api/site/config:
    get:
      summary: Get site configuration (branding and defaults)
      operationId: getSiteConfig
      tags:
        - Site
      description: >-
        Returns site-level branding and default payment settings used by the
        portal. May be public/read-only depending on environment; APIM
        enforcement can be enabled in production. If APIM is enabled, rate
        limiting headers may be present. Requests may still be required to
        traverse AFD in protected environments; direct APIM origin may be
        rejected with 403.



        Requires API key. Agents must use /api/x402/subscribe to pay and provision a key.
      security:
        - apimKey: []
      responses:
        "200":
          description: Site configuration
          headers:
            X-RateLimit-Limit:
              $ref: "#/components/headers/X-RateLimit-Limit"
            X-RateLimit-Remaining:
              $ref: "#/components/headers/X-RateLimit-Remaining"
            X-RateLimit-Reset:
              $ref: "#/components/headers/X-RateLimit-Reset"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SiteConfig"
              examples:
                default:
                  value:
                    theme:
                      primaryColor: "#1f2937"
                      secondaryColor: "#F54029"
                      brandLogoUrl: /cblogod.png
                      brandFaviconUrl: /favicon-32x32.png
                      brandName: BasaltSurge
                      fontFamily: Inter, ui-sans-serif, system-ui
                      receiptBackgroundUrl: /manifest.webmanifest
                    defaultPaymentToken: ETH
                    processingFeePct: 1
        "429":
          $ref: "#/components/responses/TooManyRequests"
...
components:
  securitySchemes:
    apimKey:
      type: apiKey
      in: header
      name: Ocp-Apim-Subscription-Key
      description: Azure API Management subscription key. Required for
        developer-facing endpoints. Agentic AI clients MUST first provision an
        API key by completing an x402 payment challenge at /api/x402/subscribe.
      x-subscribe-url: /api/x402/subscribe
    x402:
      type: apiKey
      in: header
      name: x-payment
      description: x402 Agentic Payment Standard Proof (Base64 payload)
    siwx:
      type: http
      scheme: bearer
      bearerFormat: siwx
      description: Sign-In with X (SIWX) agent authentication
  headers:
    X-RateLimit-Limit:
      description: Maximum number of requests allowed in the current window
      schema:
        type: integer
        format: int32
    X-RateLimit-Remaining:
      description: Number of requests remaining in the current window
      schema:
        type: integer
        format: int32
...